+++ context/AuthContext.tsx
@@ -1,11 +1,13 @@
 'use client'
 
 import React, { createContext, useContext, useState, useEffect } from 'react';
+import { getApolloClient} from '@/lib/graphql/apolloClient';
 import logECS from '@/utils/clientLogger';
 
 interface AuthContextType {
   isAuthenticated: boolean | null;
   setIsAuthenticated: (auth: boolean) => void;
+  clearAuthData: () => Promise<void>;
 }
 
 // Create a context with a default value
@@ -34,11 +36,35 @@ export function AuthProvider({ children }: {
       }
     };
     checkAuth();
-  }, [isAuthenticated]);
+  }, []);
+
+    // Function to clear all auth-related data
+  const clearAuthData = async () => {
+    try {
+      const apolloClient = getApolloClient();
+      if (!apolloClient) {
+        throw new Error('Apollo Client is not available');
+      }
+    
+      await apolloClient.clearStore(); // Clear Apollo cache
+      setIsAuthenticated(false);
+      logECS('info', 'Auth data and Apollo cache cleared', {
+        source: 'AuthProvider.clearAuthData'
+      });
+    } catch (err) {
+      logECS('error', `Error clearing auth data: ${err}`, {
+        source: 'AuthProvider.clearAuthData'
+      });
+    }
+  };
 
 
   return (
-    <AuthContext.Provider value={{ isAuthenticated, setIsAuthenticated }}>
+    <AuthContext.Provider value={{ 
+      isAuthenticated, 
+      setIsAuthenticated, 
+      clearAuthData 
+    }}>
       {children}
     </AuthContext.Provider>
   )